Made a class for all the file types that im supposed to do.
For jpegs i found what has to be looked for in the file in terms of headers and footers. - need to actually make it compare and search for the headers so thats what i will be doing

For gifs i also found what has to be looked for in the file in terms of headers and footers and made a block similar to the png example that will made. - need to actually make it compare and search for the headers so thats what i will be doing

For Rar files i didnt do didlly yet cause they are confusing. - need to do more research

i have ten days to figure this out probably will manage like 65% sure rn <3 .  also there is a snake *wink wink* 🐍